requirements.txt
torch
torchvision
Pillow
numpy
tk